+++ src/utils/external-resources.ts
