+++ src/utils/variable.test.ts
